Check multiple DBs with one key
The flag -dbs was added to make it possible to specify only one key for -k in hdbuserstore for the SystemDB and then specify to cleanup multiple tenants in that system. It is a prerequisite that all tenants include in that cleanup has to have a DB user with the same credentials as the one defined in the hdbuserstore for the SystemDB. See new slide 18-19 in hanacleaner.pdf for more info. Thanks to Axel Rehe for this suggestion!