Ignore Checks, also for logs
With the new -il flag, you can specify what mini-checks you want to totally ignore, i.e. even if they are critical and even if they are being watched by a check-mapping flag (e.g. the catch all flag, -ca) they will not even be mentioned in the log files. (The -ic flag only ignores the check for the check all flag, -ca, and still mentions the potential critical check in the log files, if it is mapped, e.g. with -ca.)